backend/requirements.txt
flask
flask-cors
playwright
python-dotenv
google-generativeai